extract index selector functions as composable
