+++ packages/vlossom/src/composables/index.ts
@@ -2,6 +2,7 @@ export * from './anchor-positioning-composable';
 export * from './color-scheme-composable';
 export * from './esc-close-composable';
 export * from './form-provide-composable';
+export * from './index-selector-composable';
 export * from './input-composable';
 export * from './input-form-composable';
 export * from './input-modifier-composable';
