+++ packages/vlossom/src/composables/useIndexSelector.ts
@@ -1,0 +1,5 @@
+import { Ref } from 'vue';
+
+export function useIndexSelector(list: Ref<string[]>) {
+    const index = ref(0);
+}
